+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Add `use_field_init_shorthand` config option.
 - Add `reorder_modules` configuration option.
 
 ## [0.3.6] 2018-01-18